♻️ refactor: Header 로직을 useHeader 훅으로 분리
